% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwb.tenders-package.R
\docType{package}
\name{kwb.tenders-package}
\alias{kwb.tenders}
\alias{kwb.tenders-package}
\title{kwb.tenders: R Package for Automated Monitoring of German Public Procurement Portals (Vergabeportale) for KWB-Relevant Tenders}
\description{
Logs into public procurement portals (starting with Vergabemarktplatz Brandenburg), scrapes published tenders, scores them for relevance to KWB research topics (e.g. groundwater) and renders an overview report.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KWB-R/kwb.tenders}
  \item Report bugs at \url{https://github.com/KWB-R/kwb.tenders/issues}
}

}
\author{
\strong{Maintainer}: Michael Rustler \email{michael.rustler@kompetenz-wasser.de} (\href{https://orcid.org/0000-0003-0647-7726}{ORCID})

Authors:
\itemize{
  \item Michael Rustler \email{michael.rustler@kompetenz-wasser.de} (\href{https://orcid.org/0000-0003-0647-7726}{ORCID})
}

Other contributors:
\itemize{
  \item Kompetenzzentrum Wasser Berlin gGmbH (KWB) [copyright holder]
}

}
\keyword{internal}
