% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detail.R
\name{tender_cpv_map}
\alias{tender_cpv_map}
\title{CPV-code to research-group mapping}
\usage{
tender_cpv_map(
  path = system.file("extdata", "cpv_groups.yml", package = "kwb.tenders")
)
}
\arguments{
\item{path}{YAML file mapping CPV prefixes to group slugs
(\code{inst/extdata/cpv_groups.yml}).}
}
\value{
A list of entries, each a list with \code{prefix} and \code{groups}.
}
\description{
CPV-code to research-group mapping
}
\examples{
str(tender_cpv_map())
}
